refactor: update nav bar
